+++ web/viewer-geckoview.html
@@ -60,7 +60,6 @@ See https://github.com/adobe-type-tools/cmap-resources
           "web-pdf_outline_viewer": "./stubs-geckoview.js",
           "web-pdf_presentation_mode": "./stubs-geckoview.js",
           "web-pdf_sidebar": "./stubs-geckoview.js",
-          "web-pdf_sidebar_resizer": "./stubs-geckoview.js",
           "web-pdf_thumbnail_viewer": "./stubs-geckoview.js",
           "web-print_service": "./stubs-geckoview.js",
           "web-secondary_toolbar": "./stubs-geckoview.js",
