+++ web/viewer.html
@@ -71,7 +71,6 @@ See https://github.com/adobe-type-tools/cmap-resources
           "web-pdf_outline_viewer": "./pdf_outline_viewer.js",
           "web-pdf_presentation_mode": "./pdf_presentation_mode.js",
           "web-pdf_sidebar": "./pdf_sidebar.js",
-          "web-pdf_sidebar_resizer": "./pdf_sidebar_resizer.js",
           "web-pdf_thumbnail_viewer": "./pdf_thumbnail_viewer.js",
           "web-print_service": "./pdf_print_service.js",
           "web-secondary_toolbar": "./secondary_toolbar.js",
